fix(manifest): address PR review feedback from mfaferek93
- Extract ScriptEntryConfig/ScriptsConfig to lightweight script_types.hpp
  to avoid pulling heavy default_script_provider.hpp into manifest.hpp
- Use recursive YAML-to-JSON conversion (JsonSerializer::yaml_to_json)
  for parameters_schema - nested objects and arrays now fully preserved
- Guard env parsing against non-scalar YAML values (skip instead of crash)
- Cache get_scripts() result in gateway_node to avoid redundant copy
- Use else branch instead of redundant if(empty) for scripts_dir warning
- Add #include <algorithm> for std::max
- Update docs: parameters_schema now fully supports nested structures
- Update test: verify nested schema objects/arrays round-trip correctly
